Format: Cleanup 42 login logic && front end login
