+++ static/shared/refresh_token.js
@@ -10,22 +10,19 @@ export async function refreshFetch(url, fetchBody) {
 		const response = await fetch(url, fetchBody);
 
 		if (response.status == 401) {
-			const formData = new FormData();
-			formData.append("csrfmiddlewaretoken", getCookie("csrftoken"));
-			formData.append("refresh", "");
-
 			const refresh = await fetch(global.fetch.refreshURL, {
 				method: "POST",
-				body: formData,
 			});
 
-			if (refresh.status == 200) {
+			if (refresh.status == 401) {
+				global.ui.auth = 0;
+				windowResize();
+			}
+			else if (refresh.status == 200) {
 				global.ui.auth = 1;
 				const retry = await fetch(url, fetchBody);
 				return (retry);
 			} else {
-				global.ui.auth = 0;
-				windowResize();
 				return (refresh);
 			}
 		}
