+++ templates/index.html
@@ -48,7 +48,7 @@
 							</div>
 							<hr>
 							<div class="foutytwo-link">
-								<a href="#" id="start"><img src="{% static '42Logo.png'%}"
+								<a href="#" id="login-42"><img src="{% static '42Logo.png'%}"
 										style="width:42px;height:42px;border-radius:10px;"></a>
 								<button type="button" class="login-local"><i class="fas fa-location-dot fa-2x"></i></button>
 							</div>
